reorg test files only, ported from https://github.com/open-telemetry/opentelemetry-log-collection/pull/165
