+++ operator/builtin/input/file/file_test.go
@@ -3,251 +3,18 @@ package file
 import (
 	"context"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"log"
-	"math/rand"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strconv"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/observiq/nanojack"
 	"github.com/observiq/stanza/entry"
-	"github.com/observiq/stanza/operator"
 	"github.com/observiq/stanza/operator/helper"
 	"github.com/observiq/stanza/testutil"
 	"github.com/stretchr/testify/require"
 )
-
-func newDefaultConfig(tempDir string) *InputConfig {
-	cfg := NewInputConfig("testfile")
-	cfg.PollInterval = helper.Duration{Duration: 50 * time.Millisecond}
-	cfg.StartAt = "beginning"
-	cfg.Include = []string{fmt.Sprintf("%s/*", tempDir)}
-	cfg.OutputIDs = []string{"fake"}
-	return cfg
-}
-
-func newTestFileOperator(t *testing.T, cfgMod func(*InputConfig), outMod func(*testutil.FakeOutput)) (*InputOperator, chan *entry.Entry, string) {
-	fakeOutput := testutil.NewFakeOutput(t)
-	if outMod != nil {
-		outMod(fakeOutput)
-	}
-
-	tempDir := testutil.NewTempDir(t)
-
-	cfg := newDefaultConfig(tempDir)
-	if cfgMod != nil {
-		cfgMod(cfg)
-	}
-	ops, err := cfg.Build(testutil.NewBuildContext(t))
-	require.NoError(t, err)
-	op := ops[0]
-
-	err = op.SetOutputs([]operator.Operator{fakeOutput})
-	require.NoError(t, err)
-
-	return op.(*InputOperator), fakeOutput.Received, tempDir
-}
-
-func openFile(t testing.TB, path string) *os.File {
-	file, err := os.OpenFile(path, os.O_CREATE|os.O_RDWR, 0777)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = file.Close() })
-	return file
-}
-
-func openTemp(t testing.TB, tempDir string) *os.File {
-	return openTempWithPattern(t, tempDir, "")
-}
-
-func reopenTemp(t testing.TB, name string) *os.File {
-	return openTempWithPattern(t, filepath.Dir(name), filepath.Base(name))
-}
-
-func openTempWithPattern(t testing.TB, tempDir, pattern string) *os.File {
-	file, err := ioutil.TempFile(tempDir, pattern)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = file.Close() })
-	return file
-}
-
-func getRotatingLogger(t testing.TB, tempDir string, maxLines, maxBackups int, copyTruncate, sequential bool) *log.Logger {
-	file, err := ioutil.TempFile(tempDir, "")
-	require.NoError(t, err)
-	require.NoError(t, file.Close()) // will be managed by rotator
-
-	rotator := nanojack.Logger{
-		Filename:     file.Name(),
-		MaxLines:     maxLines,
-		MaxBackups:   maxBackups,
-		CopyTruncate: copyTruncate,
-		Sequential:   sequential,
-	}
-
-	t.Cleanup(func() { _ = rotator.Close() })
-
-	return log.New(&rotator, "", 0)
-}
-
-func writeString(t testing.TB, file *os.File, s string) {
-	_, err := file.WriteString(s)
-	require.NoError(t, err)
-}
-
-func TestBuild(t *testing.T) {
-	t.Parallel()
-	fakeOutput := testutil.NewMockOperator("$.fake")
-
-	basicConfig := func() *InputConfig {
-		cfg := NewInputConfig("testfile")
-		cfg.OutputIDs = []string{"fake"}
-		cfg.Include = []string{"/var/log/testpath.*"}
-		cfg.Exclude = []string{"/var/log/testpath.ex*"}
-		cfg.PollInterval = helper.Duration{Duration: 10 * time.Millisecond}
-		return cfg
-	}
-
-	cases := []struct {
-		name             string
-		modifyBaseConfig func(*InputConfig)
-		errorRequirement require.ErrorAssertionFunc
-		validate         func(*testing.T, *InputOperator)
-	}{
-		{
-			"Basic",
-			func(f *InputConfig) { return },
-			require.NoError,
-			func(t *testing.T, f *InputOperator) {
-				require.Equal(t, f.OutputOperators[0], fakeOutput)
-				require.Equal(t, f.Include, []string{"/var/log/testpath.*"})
-				require.Equal(t, f.FilePathField, entry.NewNilField())
-				require.Equal(t, f.FileNameField, entry.NewLabelField("file_name"))
-				require.Equal(t, f.PollInterval, 10*time.Millisecond)
-			},
-		},
-		{
-			"BadIncludeGlob",
-			func(f *InputConfig) {
-				f.Include = []string{"["}
-			},
-			require.Error,
-			nil,
-		},
-		{
-			"BadExcludeGlob",
-			func(f *InputConfig) {
-				f.Include = []string{"["}
-			},
-			require.Error,
-			nil,
-		},
-		{
-			"MultilineConfiguredStartAndEndPatterns",
-			func(f *InputConfig) {
-				f.Multiline = helper.MultilineConfig{
-					LineEndPattern:   "Exists",
-					LineStartPattern: "Exists",
-				}
-			},
-			require.Error,
-			nil,
-		},
-		{
-			"MultilineConfiguredStartPattern",
-			func(f *InputConfig) {
-				f.Multiline = helper.MultilineConfig{
-					LineStartPattern: "START.*",
-				}
-			},
-			require.NoError,
-			func(t *testing.T, f *InputOperator) {},
-		},
-		{
-			"MultilineConfiguredEndPattern",
-			func(f *InputConfig) {
-				f.Multiline = helper.MultilineConfig{
-					LineEndPattern: "END.*",
-				}
-			},
-			require.NoError,
-			func(t *testing.T, f *InputOperator) {},
-		},
-		{
-			"InvalidEncoding",
-			func(f *InputConfig) {
-				f.Encoding = helper.EncodingConfig{Encoding: "UTF-3233"}
-			},
-			require.Error,
-			nil,
-		},
-		{
-			"LineStartAndEnd",
-			func(f *InputConfig) {
-				f.Multiline = helper.MultilineConfig{
-					LineStartPattern: ".*",
-					LineEndPattern:   ".*",
-				}
-			},
-			require.Error,
-			nil,
-		},
-		{
-			"NoLineStartOrEnd",
-			func(f *InputConfig) {
-				f.Multiline = helper.MultilineConfig{}
-			},
-			require.NoError,
-			func(t *testing.T, f *InputOperator) {},
-		},
-		{
-			"InvalidLineStartRegex",
-			func(f *InputConfig) {
-				f.Multiline = helper.MultilineConfig{
-					LineStartPattern: "(",
-				}
-			},
-			require.Error,
-			nil,
-		},
-		{
-			"InvalidLineEndRegex",
-			func(f *InputConfig) {
-				f.Multiline = helper.MultilineConfig{
-					LineEndPattern: "(",
-				}
-			},
-			require.Error,
-			nil,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc := tc
-			t.Parallel()
-			cfg := basicConfig()
-			tc.modifyBaseConfig(cfg)
-
-			ops, err := cfg.Build(testutil.NewBuildContext(t))
-			tc.errorRequirement(t, err)
-			if err != nil {
-				return
-			}
-			op := ops[0]
-
-			err = op.SetOutputs([]operator.Operator{fakeOutput})
-			require.NoError(t, err)
-
-			fileInput := op.(*InputOperator)
-			tc.validate(t, fileInput)
-		})
-	}
-}
 
 func TestCleanStop(t *testing.T) {
 	t.Parallel()
@@ -536,421 +303,6 @@ func TestMultiFileParallel_LiveFiles(t *testing.T) {
 	wg.Wait()
 }
 
-func TestMultiFileRotate(t *testing.T) {
-	t.Parallel()
-
-	getMessage := func(f, k, m int) string { return fmt.Sprintf("file %d-%d, message %d", f, k, m) }
-
-	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-
-	numFiles := 3
-	numMessages := 3
-	numRotations := 3
-
-	expected := make([]string, 0, numFiles*numMessages*numRotations)
-	for i := 0; i < numFiles; i++ {
-		for j := 0; j < numMessages; j++ {
-			for k := 0; k < numRotations; k++ {
-				expected = append(expected, getMessage(i, k, j))
-			}
-		}
-	}
-
-	require.NoError(t, operator.Start())
-	defer operator.Stop()
-
-	temps := make([]*os.File, 0, numFiles)
-	for i := 0; i < numFiles; i++ {
-		temps = append(temps, openTemp(t, tempDir))
-	}
-
-	var wg sync.WaitGroup
-	for i, temp := range temps {
-		wg.Add(1)
-		go func(tf *os.File, f int) {
-			defer wg.Done()
-			for k := 0; k < numRotations; k++ {
-				for j := 0; j < numMessages; j++ {
-					writeString(t, tf, getMessage(f, k, j)+"\n")
-				}
-
-				require.NoError(t, tf.Close())
-				require.NoError(t, os.Rename(tf.Name(), fmt.Sprintf("%s.%d", tf.Name(), k)))
-				tf = reopenTemp(t, tf.Name())
-			}
-		}(temp, i)
-	}
-
-	waitForMessages(t, logReceived, expected)
-	wg.Wait()
-}
-
-func TestMultiFileRotateSlow(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		// Windows has very poor support for moving active files, so rotation is less commonly used
-		// This may possibly be handled better in Go 1.16: https://github.com/golang/go/issues/35358
-		t.Skip()
-	}
-
-	t.Parallel()
-
-	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-
-	getMessage := func(f, k, m int) string { return fmt.Sprintf("file %d-%d, message %d", f, k, m) }
-	fileName := func(f, k int) string { return filepath.Join(tempDir, fmt.Sprintf("file%d.rot%d.log", f, k)) }
-	baseFileName := func(f int) string { return filepath.Join(tempDir, fmt.Sprintf("file%d.log", f)) }
-
-	numFiles := 3
-	numMessages := 30
-	numRotations := 3
-
-	expected := make([]string, 0, numFiles*numMessages*numRotations)
-	for i := 0; i < numFiles; i++ {
-		for j := 0; j < numMessages; j++ {
-			for k := 0; k < numRotations; k++ {
-				expected = append(expected, getMessage(i, k, j))
-			}
-		}
-	}
-
-	require.NoError(t, operator.Start())
-	defer operator.Stop()
-
-	var wg sync.WaitGroup
-	for fileNum := 0; fileNum < numFiles; fileNum++ {
-		wg.Add(1)
-		go func(fn int) {
-			defer wg.Done()
-
-			for rotationNum := 0; rotationNum < numRotations; rotationNum++ {
-				file := openFile(t, baseFileName(fn))
-				for messageNum := 0; messageNum < numMessages; messageNum++ {
-					writeString(t, file, getMessage(fn, rotationNum, messageNum)+"\n")
-					time.Sleep(5 * time.Millisecond)
-				}
-
-				require.NoError(t, file.Close())
-				require.NoError(t, os.Rename(baseFileName(fn), fileName(fn, rotationNum)))
-			}
-		}(fileNum)
-	}
-
-	waitForMessages(t, logReceived, expected)
-	wg.Wait()
-}
-
-func TestMultiCopyTruncateSlow(t *testing.T) {
-	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-
-	getMessage := func(f, k, m int) string { return fmt.Sprintf("file %d-%d, message %d", f, k, m) }
-	fileName := func(f, k int) string { return filepath.Join(tempDir, fmt.Sprintf("file%d.rot%d.log", f, k)) }
-	baseFileName := func(f int) string { return filepath.Join(tempDir, fmt.Sprintf("file%d.log", f)) }
-
-	numFiles := 3
-	numMessages := 30
-	numRotations := 3
-
-	expected := make([]string, 0, numFiles*numMessages*numRotations)
-	for i := 0; i < numFiles; i++ {
-		for j := 0; j < numMessages; j++ {
-			for k := 0; k < numRotations; k++ {
-				expected = append(expected, getMessage(i, k, j))
-			}
-		}
-	}
-
-	require.NoError(t, operator.Start())
-	defer operator.Stop()
-
-	var wg sync.WaitGroup
-	for fileNum := 0; fileNum < numFiles; fileNum++ {
-		wg.Add(1)
-		go func(fn int) {
-			defer wg.Done()
-
-			for rotationNum := 0; rotationNum < numRotations; rotationNum++ {
-				file := openFile(t, baseFileName(fn))
-				for messageNum := 0; messageNum < numMessages; messageNum++ {
-					writeString(t, file, getMessage(fn, rotationNum, messageNum)+"\n")
-					time.Sleep(5 * time.Millisecond)
-				}
-
-				_, err := file.Seek(0, 0)
-				require.NoError(t, err)
-				dst := openFile(t, fileName(fn, rotationNum))
-				_, err = io.Copy(dst, file)
-				require.NoError(t, err)
-				require.NoError(t, dst.Close())
-				require.NoError(t, file.Truncate(0))
-				_, err = file.Seek(0, 0)
-				require.NoError(t, err)
-				file.Close()
-			}
-		}(fileNum)
-	}
-
-	waitForMessages(t, logReceived, expected)
-	wg.Wait()
-}
-
-type rotationTest struct {
-	name            string
-	totalLines      int
-	maxLinesPerFile int
-	maxBackupFiles  int
-	writeInterval   time.Duration
-	pollInterval    time.Duration
-	ephemeralLines  bool
-}
-
-/*
-	When log files are rotated at extreme speeds, it is possible to miss some log entries.
-	This can happen when an individual log entry is written and deleted within the duration
-	of a single poll interval. For example, consider the following scenario:
-		- A log file may have up to 9 backups (10 total log files)
-		- Each log file may contain up to 10 entries
-		- Log entries are written at an interval of 10µs
-		- Log files are polled at an interval of 100ms
-	In this scenario, a log entry that is written may only exist on disk for about 1ms.
-	A polling interval of 100ms will most likely never produce a chance to read the log file.
-
-	In production settings, this consideration is not very likely to be a problem, but it is
-	easy to encounter the issue in tests, and difficult to deterministically simulate edge cases.
-	However, the above understanding does allow for some consistent expectations.
-		1) Cases that do not require deletion of old log entries should always pass.
-		2) Cases where the polling interval is sufficiently rapid should always pass.
-		3) When neither 1 nor 2 is true, there may be missing entries, but still no duplicates.
-
-	The following method is provided largely as documentation of how this is expected to behave.
-	In practice, timing is largely dependent on the responsiveness of system calls.
-*/
-func (rt rotationTest) expectEphemeralLines() bool {
-	// primary + backups
-	maxLinesInAllFiles := rt.maxLinesPerFile + rt.maxLinesPerFile*rt.maxBackupFiles
-
-	// Will the test write enough lines to result in deletion of oldest backups?
-	maxBackupsExceeded := rt.totalLines > maxLinesInAllFiles
-
-	// last line written in primary file will exist for l*b more writes
-	minTimeToLive := time.Duration(int(rt.writeInterval) * rt.maxLinesPerFile * rt.maxBackupFiles)
-
-	// can a line be written and then rotated to deletion before ever observed?
-	return maxBackupsExceeded && rt.pollInterval > minTimeToLive
-}
-
-func (rt rotationTest) run(tc rotationTest, copyTruncate, sequential bool) func(t *testing.T) {
-	return func(t *testing.T) {
-		operator, logReceived, tempDir := newTestFileOperator(t,
-			func(cfg *InputConfig) {
-				cfg.PollInterval = helper.NewDuration(tc.pollInterval)
-			},
-			func(out *testutil.FakeOutput) {
-				out.Received = make(chan *entry.Entry, tc.totalLines)
-			},
-		)
-		logger := getRotatingLogger(t, tempDir, tc.maxLinesPerFile, tc.maxBackupFiles, copyTruncate, sequential)
-
-		expected := make([]string, 0, tc.totalLines)
-		baseStr := stringWithLength(46) // + ' 123'
-		for i := 0; i < tc.totalLines; i++ {
-			expected = append(expected, fmt.Sprintf("%s %3d", baseStr, i))
-		}
-
-		require.NoError(t, operator.Start())
-		defer operator.Stop()
-
-		for _, message := range expected {
-			logger.Println(message)
-			time.Sleep(tc.writeInterval)
-		}
-
-		received := make([]string, 0, tc.totalLines)
-	LOOP:
-		for {
-			select {
-			case e := <-logReceived:
-				received = append(received, e.Record.(string))
-			case <-time.After(100 * time.Millisecond):
-				break LOOP
-			}
-		}
-
-		if tc.ephemeralLines {
-			if !tc.expectEphemeralLines() {
-				// This is helpful for test development, and ensures the sample computation is used
-				t.Logf("Potentially unstable ephemerality expectation for test: %s", tc.name)
-			}
-			require.Subset(t, expected, received)
-		} else {
-			require.ElementsMatch(t, expected, received)
-		}
-	}
-}
-
-func TestRotation(t *testing.T) {
-
-	cases := []rotationTest{
-		{
-			name:            "Fast/NoRotation",
-			totalLines:      10,
-			maxLinesPerFile: 10,
-			maxBackupFiles:  1,
-			writeInterval:   time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-		},
-		{
-			name:            "Fast/NoDeletion",
-			totalLines:      20,
-			maxLinesPerFile: 10,
-			maxBackupFiles:  1,
-			writeInterval:   time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-		},
-		{
-			name:            "Fast/Deletion",
-			totalLines:      30,
-			maxLinesPerFile: 10,
-			maxBackupFiles:  1,
-			writeInterval:   time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-			ephemeralLines:  true,
-		},
-		{
-			name:            "Fast/Deletion/ExceedFingerprint",
-			totalLines:      300,
-			maxLinesPerFile: 100,
-			maxBackupFiles:  1,
-			writeInterval:   time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-			ephemeralLines:  true,
-		},
-		{
-			name:            "Slow/NoRotation",
-			totalLines:      10,
-			maxLinesPerFile: 10,
-			maxBackupFiles:  1,
-			writeInterval:   3 * time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-		},
-		{
-			name:            "Slow/NoDeletion",
-			totalLines:      20,
-			maxLinesPerFile: 10,
-			maxBackupFiles:  1,
-			writeInterval:   3 * time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-		},
-		{
-			name:            "Slow/Deletion",
-			totalLines:      30,
-			maxLinesPerFile: 10,
-			maxBackupFiles:  1,
-			writeInterval:   3 * time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-		},
-		{
-			name:            "Slow/Deletion/ExceedFingerprint",
-			totalLines:      100,
-			maxLinesPerFile: 25, // ~20 is just enough to exceed 1000 bytes fingerprint at 50 chars per line
-			maxBackupFiles:  2,
-			writeInterval:   3 * time.Millisecond,
-			pollInterval:    20 * time.Millisecond,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(fmt.Sprintf("%s/MoveCreateTimestamped", tc.name), tc.run(tc, false, false))
-		t.Run(fmt.Sprintf("%s/MoveCreateSequential", tc.name), tc.run(tc, false, true))
-		t.Run(fmt.Sprintf("%s/CopyTruncateTimestamped", tc.name), tc.run(tc, true, false))
-		t.Run(fmt.Sprintf("%s/CopyTruncateSequential", tc.name), tc.run(tc, true, true))
-	}
-}
-
-func TestMoveFile(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		t.Skip("Moving files while open is unsupported on Windows")
-	}
-	t.Parallel()
-	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-
-	temp1 := openTemp(t, tempDir)
-	writeString(t, temp1, "testlog1\n")
-	temp1.Close()
-
-	operator.poll(context.Background())
-	defer operator.Stop()
-
-	waitForMessage(t, logReceived, "testlog1")
-
-	// Wait until all goroutines are finished before renaming
-	operator.wg.Wait()
-	err := os.Rename(temp1.Name(), fmt.Sprintf("%s.2", temp1.Name()))
-	require.NoError(t, err)
-
-	operator.poll(context.Background())
-	expectNoMessages(t, logReceived)
-}
-
-// TruncateThenWrite tests that, after a file has been truncated,
-// any new writes are picked up
-func TestTruncateThenWrite(t *testing.T) {
-	t.Parallel()
-	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-
-	temp1 := openTemp(t, tempDir)
-	writeString(t, temp1, "testlog1\ntestlog2\n")
-
-	operator.poll(context.Background())
-	defer operator.Stop()
-
-	waitForMessage(t, logReceived, "testlog1")
-	waitForMessage(t, logReceived, "testlog2")
-
-	require.NoError(t, temp1.Truncate(0))
-	temp1.Seek(0, 0)
-
-	writeString(t, temp1, "testlog3\n")
-	operator.poll(context.Background())
-	waitForMessage(t, logReceived, "testlog3")
-	expectNoMessages(t, logReceived)
-}
-
-// CopyTruncateWriteBoth tests that when a file is copied
-// with unread logs on the end, then the original is truncated,
-// we get the unread logs on the copy as well as any new logs
-// written to the truncated file
-func TestCopyTruncateWriteBoth(t *testing.T) {
-	t.Parallel()
-	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-
-	temp1 := openTemp(t, tempDir)
-	writeString(t, temp1, "testlog1\ntestlog2\n")
-
-	operator.poll(context.Background())
-	defer operator.Stop()
-
-	waitForMessage(t, logReceived, "testlog1")
-	waitForMessage(t, logReceived, "testlog2")
-	operator.wg.Wait() // wait for all goroutines to finish
-
-	// Copy the first file to a new file, and add another log
-	temp2 := openTemp(t, tempDir)
-	_, err := io.Copy(temp2, temp1)
-	require.NoError(t, err)
-
-	// Truncate original file
-	require.NoError(t, temp1.Truncate(0))
-	temp1.Seek(0, 0)
-
-	// Write to original and new file
-	writeString(t, temp2, "testlog3\n")
-	writeString(t, temp1, "testlog4\n")
-
-	// Expect both messages to come through
-	operator.poll(context.Background())
-	waitForMessages(t, logReceived, []string{"testlog3", "testlog4"})
-}
-
 // OffsetsAfterRestart tests that a operator is able to load
 // its offsets after a restart
 func TestOffsetsAfterRestart(t *testing.T) {
@@ -1018,37 +370,6 @@ func TestOffsetsAfterRestart_BigFilesWrittenWhileOff(t *testing.T) {
 	writeString(t, temp, log2+"\n")
 
 	// Start the operator and expect the message
-	require.NoError(t, operator.Start())
-	waitForMessage(t, logReceived, log2)
-}
-
-func TestFileMovedWhileOff_BigFiles(t *testing.T) {
-	t.Parallel()
-	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-
-	log1 := stringWithLength(1000)
-	log2 := stringWithLength(1000)
-
-	temp := openTemp(t, tempDir)
-	writeString(t, temp, log1+"\n")
-	require.NoError(t, temp.Close())
-
-	// Start the operator
-	require.NoError(t, operator.Start())
-	defer operator.Stop()
-	waitForMessage(t, logReceived, log1)
-
-	// Stop the operator, then rename and write a new log
-	require.NoError(t, operator.Stop())
-
-	err := os.Rename(temp.Name(), fmt.Sprintf("%s2", temp.Name()))
-	require.NoError(t, err)
-
-	temp = reopenTemp(t, temp.Name())
-	require.NoError(t, err)
-	writeString(t, temp, log2+"\n")
-
-	// Expect the message written to the new log to come through
 	require.NoError(t, operator.Start())
 	waitForMessage(t, logReceived, log2)
 }
@@ -1162,75 +483,6 @@ func TestFileReader_FingerprintUpdated(t *testing.T) {
 	require.Equal(t, []byte("testlog1\n"), reader.Fingerprint.FirstBytes)
 }
 
-func stringWithLength(length int) string {
-	charset := "abcdefghijklmnopqrstuvwxyz"
-	b := make([]byte, length)
-	for i := range b {
-		b[i] = charset[rand.Intn(len(charset))]
-	}
-	return string(b)
-}
-
-func waitForOne(t *testing.T, c chan *entry.Entry) *entry.Entry {
-	select {
-	case e := <-c:
-		return e
-	case <-time.After(time.Second):
-		require.FailNow(t, "Timed out waiting for message")
-		return nil
-	}
-}
-
-func waitForN(t *testing.T, c chan *entry.Entry, n int) []string {
-	messages := make([]string, 0, n)
-	for i := 0; i < n; i++ {
-		select {
-		case e := <-c:
-			messages = append(messages, e.Record.(string))
-		case <-time.After(time.Second):
-			require.FailNow(t, "Timed out waiting for message")
-			return nil
-		}
-	}
-	return messages
-}
-
-func waitForMessage(t *testing.T, c chan *entry.Entry, expected string) {
-	select {
-	case e := <-c:
-		require.Equal(t, expected, e.Record.(string))
-	case <-time.After(time.Second):
-		require.FailNow(t, "Timed out waiting for message", expected)
-	}
-}
-
-func waitForMessages(t *testing.T, c chan *entry.Entry, expected []string) {
-	receivedMessages := make([]string, 0, len(expected))
-LOOP:
-	for {
-		select {
-		case e := <-c:
-			receivedMessages = append(receivedMessages, e.Record.(string))
-		case <-time.After(time.Second):
-			break LOOP
-		}
-	}
-
-	require.ElementsMatch(t, expected, receivedMessages)
-}
-
-func expectNoMessages(t *testing.T, c chan *entry.Entry) {
-	expectNoMessagesUntil(t, c, 200*time.Millisecond)
-}
-
-func expectNoMessagesUntil(t *testing.T, c chan *entry.Entry, d time.Duration) {
-	select {
-	case e := <-c:
-		require.FailNow(t, "Received unexpected message", "Message: %s", e.Record.(string))
-	case <-time.After(d):
-	}
-}
-
 func TestEncodings(t *testing.T) {
 	t.Parallel()
 	cases := []struct {
@@ -1316,64 +568,6 @@ func TestEncodings(t *testing.T) {
 	}
 }
 
-type fileInputBenchmark struct {
-	name   string
-	config *InputConfig
-}
-
-func BenchmarkFileInput(b *testing.B) {
-	cases := []fileInputBenchmark{
-		{
-			"Default",
-			NewInputConfig("test_id"),
-		},
-		{
-			"NoFileName",
-			func() *InputConfig {
-				cfg := NewInputConfig("test_id")
-				cfg.IncludeFileName = false
-				return cfg
-			}(),
-		},
-	}
-
-	for _, tc := range cases {
-		b.Run(tc.name, func(b *testing.B) {
-			tempDir := testutil.NewTempDir(b)
-			path := filepath.Join(tempDir, "in.log")
-
-			cfg := tc.config
-			cfg.OutputIDs = []string{"fake"}
-			cfg.Include = []string{path}
-			cfg.StartAt = "beginning"
-
-			ops, err := cfg.Build(testutil.NewBuildContext(b))
-			require.NoError(b, err)
-			op := ops[0]
-
-			fakeOutput := testutil.NewFakeOutput(b)
-			err = op.SetOutputs([]operator.Operator{fakeOutput})
-			require.NoError(b, err)
-
-			err = op.Start()
-			defer op.Stop()
-			require.NoError(b, err)
-
-			file, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE, 0666)
-			require.NoError(b, err)
-
-			for i := 0; i < b.N; i++ {
-				file.WriteString("testlog\n")
-			}
-
-			b.ResetTimer()
-			for i := 0; i < b.N; i++ {
-				<-fakeOutput.Received
-			}
-		})
-	}
-}
-
 // TestExclude tests that a log file will be excluded if it matches the
 // glob specified in the operator.
 func TestExclude(t *testing.T) {
@@ -1415,15 +609,4 @@ func TestExcludeDuplicates(t *testing.T) {
 
 	matches := getMatches(includes, excludes)
 	require.ElementsMatch(t, matches, paths[2:3])
-}
-
-// writes file with the specified file names and returns their full paths in order
-func writeTempFiles(tempDir string, names []string) []string {
-	result := make([]string, 0, len(names))
-	for _, name := range names {
-		path := filepath.Join(tempDir, name)
-		ioutil.WriteFile(path, []byte(name), 0755)
-		result = append(result, path)
-	}
-	return result
 }
