+++ operator/builtin/input/file/util_test.go
@@ -1,0 +1,175 @@
+package file
+
+import (
+	"fmt"
+	"io/ioutil"
+	"log"
+	"math/rand"
+	"os"
+	"path/filepath"
+	"testing"
+	"time"
+
+	"github.com/observiq/nanojack"
+	"github.com/observiq/stanza/entry"
+	"github.com/observiq/stanza/operator"
+	"github.com/observiq/stanza/operator/helper"
+	"github.com/observiq/stanza/testutil"
+	"github.com/stretchr/testify/require"
+)
+
+func newDefaultConfig(tempDir string) *InputConfig {
+	cfg := NewInputConfig("testfile")
+	cfg.PollInterval = helper.Duration{Duration: 50 * time.Millisecond}
+	cfg.StartAt = "beginning"
+	cfg.Include = []string{fmt.Sprintf("%s/*", tempDir)}
+	cfg.OutputIDs = []string{"fake"}
+	return cfg
+}
+
+func newTestFileOperator(t *testing.T, cfgMod func(*InputConfig), outMod func(*testutil.FakeOutput)) (*InputOperator, chan *entry.Entry, string) {
+	fakeOutput := testutil.NewFakeOutput(t)
+	if outMod != nil {
+		outMod(fakeOutput)
+	}
+
+	tempDir := testutil.NewTempDir(t)
+
+	cfg := newDefaultConfig(tempDir)
+	if cfgMod != nil {
+		cfgMod(cfg)
+	}
+	ops, err := cfg.Build(testutil.NewBuildContext(t))
+	require.NoError(t, err)
+	op := ops[0]
+
+	err = op.SetOutputs([]operator.Operator{fakeOutput})
+	require.NoError(t, err)
+
+	return op.(*InputOperator), fakeOutput.Received, tempDir
+}
+
+func openFile(t testing.TB, path string) *os.File {
+	file, err := os.OpenFile(path, os.O_CREATE|os.O_RDWR, 0777)
+	require.NoError(t, err)
+	t.Cleanup(func() { _ = file.Close() })
+	return file
+}
+
+func openTemp(t testing.TB, tempDir string) *os.File {
+	return openTempWithPattern(t, tempDir, "")
+}
+
+func reopenTemp(t testing.TB, name string) *os.File {
+	return openTempWithPattern(t, filepath.Dir(name), filepath.Base(name))
+}
+
+func openTempWithPattern(t testing.TB, tempDir, pattern string) *os.File {
+	file, err := ioutil.TempFile(tempDir, pattern)
+	require.NoError(t, err)
+	t.Cleanup(func() { _ = file.Close() })
+	return file
+}
+
+func getRotatingLogger(t testing.TB, tempDir string, maxLines, maxBackups int, copyTruncate, sequential bool) *log.Logger {
+	file, err := ioutil.TempFile(tempDir, "")
+	require.NoError(t, err)
+	require.NoError(t, file.Close()) // will be managed by rotator
+
+	rotator := nanojack.Logger{
+		Filename:     file.Name(),
+		MaxLines:     maxLines,
+		MaxBackups:   maxBackups,
+		CopyTruncate: copyTruncate,
+		Sequential:   sequential,
+	}
+
+	t.Cleanup(func() { _ = rotator.Close() })
+
+	return log.New(&rotator, "", 0)
+}
+
+func writeString(t testing.TB, file *os.File, s string) {
+	_, err := file.WriteString(s)
+	require.NoError(t, err)
+}
+
+func stringWithLength(length int) string {
+	charset := "abcdefghijklmnopqrstuvwxyz"
+	b := make([]byte, length)
+	for i := range b {
+		b[i] = charset[rand.Intn(len(charset))]
+	}
+	return string(b)
+}
+
+func waitForOne(t *testing.T, c chan *entry.Entry) *entry.Entry {
+	select {
+	case e := <-c:
+		return e
+	case <-time.After(time.Second):
+		require.FailNow(t, "Timed out waiting for message")
+		return nil
+	}
+}
+
+func waitForN(t *testing.T, c chan *entry.Entry, n int) []string {
+	messages := make([]string, 0, n)
+	for i := 0; i < n; i++ {
+		select {
+		case e := <-c:
+			messages = append(messages, e.Record.(string))
+		case <-time.After(time.Second):
+			require.FailNow(t, "Timed out waiting for message")
+			return nil
+		}
+	}
+	return messages
+}
+
+func waitForMessage(t *testing.T, c chan *entry.Entry, expected string) {
+	select {
+	case e := <-c:
+		require.Equal(t, expected, e.Record.(string))
+	case <-time.After(time.Second):
+		require.FailNow(t, "Timed out waiting for message", expected)
+	}
+}
+
+func waitForMessages(t *testing.T, c chan *entry.Entry, expected []string) {
+	receivedMessages := make([]string, 0, len(expected))
+LOOP:
+	for {
+		select {
+		case e := <-c:
+			receivedMessages = append(receivedMessages, e.Record.(string))
+		case <-time.After(time.Second):
+			break LOOP
+		}
+	}
+
+	require.ElementsMatch(t, expected, receivedMessages)
+}
+
+func expectNoMessages(t *testing.T, c chan *entry.Entry) {
+	expectNoMessagesUntil(t, c, 200*time.Millisecond)
+}
+
+func expectNoMessagesUntil(t *testing.T, c chan *entry.Entry, d time.Duration) {
+	select {
+	case e := <-c:
+		require.FailNow(t, "Received unexpected message", "Message: %s", e.Record.(string))
+	case <-time.After(d):
+	}
+}
+
+// writes file with the specified file names and returns their full paths in order
+func writeTempFiles(tempDir string, names []string) []string {
+	result := make([]string, 0, len(names))
+	for _, name := range names {
+		path := filepath.Join(tempDir, name)
+		ioutil.WriteFile(path, []byte(name), 0755)
+		result = append(result, path)
+	}
+	return result
+}
